Organize router files
